responsive ok

+++ controllers/tutoController.php
@@ -1,38 +1,36 @@
 <?php
 
- if (isset($_POST['confirmDelIntro']))
+ if (isset($_POST['confirmDelIntro']) /* && isset($_POST['deleteIntro']) */)
  {
      include_once '../configuration.php';
      $deleteContent = new u01cc_contribs();
      $deleteContent->id = htmlspecialchars($_POST['introID']);
      $deleteContentResult = $deleteContent->deleteContent();
      echo true;
-          header('location: ../views/raid.php');
-
+     header('location: ../views/tuto.php');
  }
- elseif (isset($_POST['confirmDelStep']))
+ elseif (isset($_POST['confirmDelStep']) /* && isset($_POST['deleteStep']) */)
  {
      include_once '../configuration.php';
      $deleteStepContent = new u01cc_raidSteps();
      $deleteStepContent->id = htmlspecialchars($_POST['stepId']);
      $deleteStepContentResult = $deleteStepContent->deleteStepContent();
      echo true;
-          header('location: ../views/raid.php');
-
+     header('location: ../views/tuto.php');
  }
  else
  {
-      if (isset($_POST['summary']))
+     if (isset($_POST['summary']))
      {
          include_once '../configuration.php';
 
-         $getRaid = new u01cc_raidSteps();
-         $getRaid->id_u01cc_contribs = $_POST['summary'];
-         $getRaidResult = $getRaid->getRaid();
+         $getTuto = new u01cc_raidSteps();
+         $getTuto->id_u01cc_contribs = $_POST['summary'];
+         $getTutoResult = $getTuto->getTuto();
      }
      include_once '../configuration.php';
 
-     $getTitleRaid = new u01cc_contribs();
-     $getTitleRaidResult = $getTitleRaid->getTitleRaid();
+     $getTitleTuto = new u01cc_contribs();
+     $getTitleTutoResult = $getTitleTuto->getTitleTuto();
  }
 ?>
